Refactored Document Api
